Move component to designated folder
