+++ src/applications/search/components/SearchDropdown/SearchDropdownComponent.js
