+++ src/applications/search/components/SearchDropdown/SearchDropdownStyles.scss
